Refactoring to use shell script
work in progress...

Make is a poor fit for this type of work - or at least my competence
at make isn't up to the task.  This is a change to using a shell
script to initiate the execution of other shell scripts.

In addition, the repo has been updated to use token authentication
against the API endpoint.